Add Duovero board
The Gumstix Duovero is another TI OMAP4430 Cortex-A9 based board.

There is no direct support for this board in FreeBSD-CURRENT, but
the Pandaboard is similar enough that the kernel boots. This is
what this configuration is currently using.

The board uses U-Boot (v2014.10). Duovero support for this exists
and is being used.

ubldr is also used.